requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.3
bleach==6.2.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
docopt==0.6.2
executing==2.2.0
fastjsonschema==2.21.1
Flask==2.2.2
flatbuffers==20181003210633
fonttools==4.57.0
frozenlist==1.5.0
gotrue==2.12.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.8
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
ipython==8.12.3
itsdangerous==2.2.0
jax==0.5.3
jaxlib==0.5.3
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mediapipe==0.10.15
mistune==3.1.3
ml_dtypes==0.5.1
multidict==6.4.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.2.1
pipreqs==0.5.0
platformdirs==4.3.7
pluggy==1.5.0
postgrest==1.0.1
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==4.25.6
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
pyzmq==26.4.0
realtime==2.4.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.6
stack-data==0.6.3
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.0
supafunc==0.9.4
tinycss2==1.4.0
tornado==6.4.2
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==14.2
Werkzeug==2.2.2
yarg==0.1.9
yarl==1.19.0
